--- a/src/main/java/org/apache/sysds/runtime/data/SparseBlockMCSR.java
+++ b/src/main/java/org/apache/sysds/runtime/data/SparseBlockMCSR.java
@@ -382,7 +382,7 @@ public class SparseBlockMCSR extends SparseBlock
 		if(v == 0)
 			return;
 		else if(_rows[r] == null)
-			_rows[r] = new SparseRowScalar().append(c, v);
+			_rows[r] = new SparseRowScalar(c, v);
 		else
 			_rows[r] = _rows[r].append(c, v);
 	}