This is a PR to performance measure the append of MCSR, 

```
Profiling started
Append all dense:             4.181+-  0.073 ms
Append all zero on empty:     0.198+-  0.004 ms
Append all zero on Scalar:    0.210+-  0.011 ms
Append all zero on Array:     0.230+-  0.006 ms
Append half zero on Array:    4.508+-  0.194 ms
```